refactor: separated batch and layer norm into 2 files
