+++ test/autoencoder/test_autoencoder_v2.h
@@ -1,12 +1,3 @@
-///////////////////////////////////////////////////////////////////////////////
-// File:         test_autoencoder_v2.h
-// Description:  ...
-// Authors:      Luong-Ha Nguyen & James-A. Goulet
-// Created:      March 15, 2024
-// Updated:      March 15, 2024
-// Contact:      luongha.nguyen@gmail.com & james.goulet@polymtl.ca
-// License:      This code is released under the MIT License.
-////////////////////////////////////////////////////////////////////////////////
 #pragma once
 
 int test_autoecoder_v2();
