+++ test/cross_val/cross_val.h
@@ -1,13 +1,3 @@
-///////////////////////////////////////////////////////////////////////////////
-// File:         cross_val.h
-// Description:  ...
-// Authors:      Luong-Ha Nguyen & James-A. Goulet
-// Created:      February 28, 2024
-// Updated:      February 28, 2024
-// Contact:      luongha.nguyen@gmail.com & james.goulet@polymtl.ca
-// License:      This code is released under the MIT License.
-////////////////////////////////////////////////////////////////////////////////
-
 #pragma once
 
 int cross_val_with_old_version();
