+++ test/lstm/test_lstm_v2.h
@@ -1,12 +1,3 @@
-///////////////////////////////////////////////////////////////////////////////
-// File:         test_lstm_v2.cpp
-// Description:
-// Authors:      Luong-Ha Nguyen & James-A. Goulet
-// Created:      March 27, 2024
-// Updated:      March 27, 2024
-// Contact:      luongha.nguyen@gmail.com & james.goulet@polymtl.ca
-// License:      This code is released under the MIT License.
-////////////////////////////////////////////////////////////////////////////////
 #pragma once
 
 int test_lstm_v2();
