+++ test/mha/test_mha_cpu.h
@@ -1,12 +1,3 @@
-///////////////////////////////////////////////////////////////////////////////
-// File:         test_mha_cpu.h
-// Description:  Header file for unittest for multi-head self-attention
-// Authors:      Luong-Ha Nguyen & James-A. Goulet
-// Created:      May 07, 2023
-// Updated:      May 15, 2023
-// Contact:      luongha.nguyen@gmail.com & james.goulet@polymtl.ca
-// License:      This code is released under the MIT License.
-////////////////////////////////////////////////////////////////////////////////
 #pragma once
 #include <cmath>
 #include <ctime>
