+++ test/smoother/test_smoother.h
@@ -1,14 +1,3 @@
-///////////////////////////////////////////////////////////////////////////////
-// File:         test_smoother.cpp
-// Description:  Header filer for Test smoother for a toy problem using
-// Smoothing in LSTM networks.
-// Authors : Van-Dai Vuong, Luong - Ha Nguyen& James - A.Goulet
-// Created : August 21, 2024
-// Updated:  August 21, 2024
-// Contact:  van-dai.vuong@polymtl.ca, luongha.nguyen@gmail.com &
-// james.goulet@polymtl.ca
-// License : This code is released under the MIT
-////////////////////////////////////////////////////////////////////////////////
 #pragma once
 
 int test_smoother();
